chat/requirements.txt
flask
langgraph
sentence-transformers
sqlalchemy
python-dotenv
Pillow
requests
transformers
